Now this actually is a multi-user app. Code copied from APL

+++ tasks/duration_delete.php
@@ -2,6 +2,7 @@
 include('../common.php');
 
 if(isset($QUERY['duration']) and is_numeric($QUERY['duration'])) {
+	checkDurationOwnership($QUERY['duration']);
 	$sql->execQuery("DELETE FROM Duration WHERE id=$QUERY[duration]");
 
 	showMessage("Duration deleted successfully",'tasks/edit.php');
